pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["zeeref*"]

[tool.setuptools.package-data]
"*" = ["*.html", "*.png"]

[project]
name = "ZeeRef"
version = "0.3.4-dev"
description = "A simple reference image viewer"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    { name = "Zack Gomez", email = "zack.gomez@gmail.com" },
    { name = "Rebecca Breu", email = "rebecca@rbreu.de" },
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "Pillow>=10.0.0",
    "pyQt6-Qt6>=6.7.0,<=6.7.0",
    "pyQt6>=6.7.0,<=6.7.0",
    "mistune>=3.0.0",
    "rectangle-packer>=2.0.1,<=2.0.2",
]

[project.optional-dependencies]
# Kept in sync with requirements/*.txt so uv can install tooling via extras.
test = [
    "coverage==7.5.1",
    "httpretty==1.1.4",
    "pytest==8.2.0",
    "pytest-qt==4.4.0",
    "pytest-cov==5.0.0",
]
build = [
    "pyinstaller==6.6.0",
]
dev = [
    # -r test.txt
    "coverage==7.5.1",
    "httpretty==1.1.4",
    "pytest==8.2.0",
    "pytest-qt==4.4.0",
    "pytest-cov==5.0.0",
    # -r build.txt
    "pyinstaller==6.6.0",
    # Dev-only
    "ruff>=0.4.0",
    "pybadges==3.0.1",
    "yamllint==1.35.1",
]

[project.gui-scripts]
zeeref = "zeeref.__main__:main"

[project.urls]
Repository = "https://github.com/zackgomez/zeeref"
Issues = "https://github.com/zackgomez/zeeref/issues"

[dependency-groups]
dev = [
    "ty>=0.0.23",
]

[tool.ty.rules]
invalid-method-override = "ignore"
